refactor(numbers): 화면에 타이머 추가 및 파일 이름, 게임 이름 변경
