+++ src/pages/Games/Number/components/NumberCard.tsx
@@ -1,6 +1,6 @@
 import CardBack from "./CardBack";
 import CardFront from "./CardFront";
-import { motion, AnimatePresence } from "framer-motion";
+import ShuffleCard from "./ShuffleCard";
 
 interface Props {
   cardStatus: string;
@@ -9,6 +9,7 @@ interface Props {
   gridSize: number;
   round: number;
   setUserAnswer: React.Dispatch<React.SetStateAction<number[]>>;
+  userRound: number;
 }
 
 function NumberCard({
@@ -18,32 +19,30 @@ function NumberCard({
   gridSize,
   randomNumberList,
   round,
+  userRound,
 }: Props) {
   return (
     <>
-      <AnimatePresence mode="wait">
-        {cardStatus === "front" && (
-          <motion.div key={`front-${round}`}>
-            <CardFront
-              key={round}
-              gameStep={gameStep}
-              randomNumberList={randomNumberList!}
-              gridSize={gridSize}
-            />
-          </motion.div>
-        )}
-        {cardStatus === "back" && (
-          <motion.div key={`back-${round}`}>
-            <CardBack
-              key={round}
-              gameStep={gameStep}
-              randomNumberList={randomNumberList!}
-              gridSize={gridSize}
-              setUserAnswer={setUserAnswer}
-            />
-          </motion.div>
-        )}
-      </AnimatePresence>
+      {cardStatus === "shuffle" && (
+        <ShuffleCard key={round} userRound={userRound} />
+      )}
+      {cardStatus === "front" && (
+        <CardFront
+          key={round}
+          gameStep={gameStep}
+          randomNumberList={randomNumberList!}
+          gridSize={gridSize}
+        />
+      )}
+      {cardStatus === "back" && (
+        <CardBack
+          key={round}
+          gameStep={gameStep}
+          randomNumberList={randomNumberList!}
+          gridSize={gridSize}
+          setUserAnswer={setUserAnswer}
+        />
+      )}
     </>
   );
 }
